Save functionality
Maaan, initially tried to put all the save functionality in the Game Instance BP, but wasn't working, it wouldn't load the savegame or display the player stats.  So reverted back to duplicating the save, load and checkforsavegame in all the menu widget BPs.  But, realised I was missing a SaveGameCheck in the TitleMenum so the inital GameInstance implementation will probably work now.  But I checking this fucked in whilst it still works!